First Decent Prototype! Added in audio resources and more!
REMOVED TransitionScene. This is now MainMenu. Added in all the audio from the drive thus far for ease of use. We have done great work to get this far and hopefuly this build will allow the design team to playtest effectively. That said, we still have a lot of work ahead of us but I am proud of what we have done. FYI, nothing happens if you click the dodge ball in the title.