Updated Save Manager
Save Manager now has Save( ), Read( ), ReadLevels( ), and ReadBooleans( ). Save( ) takes current character levels (and xp once that is being tracked), along with whichever conditions have been completed, and saves them. Read( ) looks at the file where data is saved, and, if there is daved data, changes the variables of the class to the stored variables. ReadLevels( ) and ReadBooleans( ) read the corresponding saved values, and return them in arrays, in case that data is wanted but without overwriting the current data.